Workaround for Kivy > 2.0.0rc3
Kivy commit 3c2a6d5 imports kivy.core.window in kivy.uix.textinput.
This is problematic because if kivy.core.window is imported in
a different thread than the main app runs in, a segmentation
fault occurs. This workaround defers the import of kivy.uix.textinput
to the kivy thread.