load-cell-probe: New module for load-cell based probes
Add support for load-cell based probes. These probes utilise a
load-cell sensor measuring the contact force between hotend and
nozzle. The provided algorithm is capable to deal with several
issues like baseline wandering and hysteresis effects which are
often inherent to such systems. It provides a true, offset-free
measurement of the bed surface without the need to calibrate the
probe.

Signed-off-by: Martin Hierholzer <martin@hierholzer.info>